+++ 4.0.0/src/BoilerPlate.Application/OsobaAppService/Dto/OsobaPutDto.cs
@@ -13,6 +13,6 @@ namespace BoilerPlate.OsobaAppService.Dto
     {
     public string Ime { get; set; }
     public string Prezime { get; set; }
-    public long KancelarijaId { get; set; }
+    public int KancelarijaId { get; set; }
     }
 }
